Setup automated CI releases (#18)
Previously, I have been releasing the plugin from my local computer
using a manual process. This was problematic because it was easy to
screw up the release by having a wrong version of Cody cloned locally,
or forgetting to publish an accompanying alpha release for every stable
release. The 3.4.0 had a critical regression that was caused by this
exact root problem.

This PR is the first step towards addressing the problem by adding CI
workflows to publish releases. By publishing the releases in CI, we
document all the necessary steps to publish a release, and we enable
anyone on the team to cut a release at any time.

By default, we only publish alpha releases for now. I want to follow up
with a separate PR that automates the process to "promote" an alpha
release to stable. My goal is that we use a byte-by-byte identical
release to the alpha when promote to stable.



## Test plan

Difficult to test this without running the actual workflows in CI.

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles

Why does it matter? 

These test plans are there to demonstrate that are following industry
standards which are important or critical for our customers.
They might be read by customers or an auditor. There are meant be simple
and easy to read. Simply explain what you did to ensure
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component: 
- "Covered by existing tests" or "CI" for the shortest possible plan if
there is zero ambiguity
  - "Added new tests" 
- "Manually tested" (if non trivial, share some output, logs, or
screenshot)
- Updating docs: 
  - "previewed locally" 
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if
incorrect
  - "CI" 
  - "locally tested" 
-->