fix: apply latency only when last suggestion was read (#1394)
RE: https://github.com/sourcegraph/cody/pull/1351

fix: apply min latency only when last suggestion was read

The minimum latency for code suggestions is now only applied when the
last suggestion was read by the user. This avoids applying the latency
when the user is actively typing and has not read the previous
suggestions.

The lastCompletionRequestTimestamp property tracks when the last
suggestion request was made. The latency is skipped if the time since
last request is less than 750ms.


## Test plan

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

Check the output channel. You should not see latency getting increased
after 5 keystrokes, which is the current behavior.