feat: feature flag for making less unhelpful suggestion requests (#1392)
REF:
https://sourcegraph.slack.com/archives/C05AGQYD528/p1697227062174209?thread_ts=1697115251.236759&cid=C05AGQYD528

feat: new feature flag for increased autocomplete debounce time

This commit is proposing to add a new feature flag called
CodyAutocompleteLowPerformanceDebounce to enable increased debounce time
for autocomplete in low-performance languages.

The debounce time is increased to 1000ms (the default number we used for
added latency) for single lines and 1500ms for multi-line (which is rare
in most of the low-performance langs) in low-performance languages when
the feature flag is enabled.

The list of low-performance language IDs is defined in
lowPerformanceLanguageIds.

This allows us to experiment with increased debounce time, specifically
for low-performance languages.

### Proposal

My rationale is that by adding latency to low-performance language, we
are still making the same amount of requests (between low-performance
languages and those that are not), we just display them with a delay. It
improves user experience without any impact.

With debounce, we will not only reduce the number of unhelpful
suggestions that contribute to costs. The users will also benefit the
same from getting fewer unhelpful suggestions, which count towards their
daily rate limit. (unless I misunderstood how latency and debounce work,
again 🤡 )




## Test plan

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

Enable the 'cody-autocomplete-low-performance-debounce' and work in a
low-performance file to check if you are still getting suggestions as
often as before.