Autocomplete: add a feature flag for streaming truncation (#1404)
- Moved the early truncation logic for LLM streaming behind the feature
flag (enabled by default).
- Moved the completion parsing logic to the provider streaming callback
so that we can later quickly enable
https://github.com/sourcegraph/cody/issues/1402, which requires the
completion text to be inserted into the document parse tree for
analysis.
- Updated unit tests accordingly to preserve all test cases.
- Changed the behavior of `trimUntilSuffix` to start from the end of
completion. It ensures that we trim only from the end completion. See
the updated unit test for a use-case `stops when the next non-empty line
of the suffix matches`.