fix: use selectionRange in edits when available (#1429)
Fix regression caused by changes in
https://github.com/sourcegraph/cody/pull/1317

The commit fixes an issue where getSmartSelection was always
recalculating a selection even when a valid selectionRange was already
provided.

Currently, the /edit command does not respect user selection when it's
available:


https://github.com/sourcegraph/cody/assets/68532117/1b07e72f-3802-4f2a-8f14-f2abbb33ae84

This becomes an issue as folding range is still in the experimental
stage. It also cause the code lens to jump around, unless it's intended
as I haven't read through the PR 😅

This PR adds a check to use selectionRange when it is available and not
an empty range.

This provides a fallback in case where "smart" selection is not working
for users.

## Test plan

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

Select code in your editor and run the /edit command, the range should
not be expanded for you.