BFG: add new graph context engine (#1232)
Previously, the "Graph Context" feature flag supported two values, `lsp-light` and `null`.
This PR adds a third option `bfg`, which is a new code graph engine that is implemented
by a separate binary application that's written in Rust. The VS Code extension download the BFG
binary from https://github.com/sourcegraph/bfg/releases, and communicates with the BFG binary
via JSON-RPC through a bespoke BFG protocol, which is defined in the file `vscode/src/jsonrpc/bfg-protocol.ts`.

## Test plan

See `agent/src/bfg/BfgContextFetcher.test.ts`. This test is disabled in CI for now since it relies on downloading a
binary from GitHub releases. We plan to enable this test in CI in a followup PR.

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->