Agent: exist process when stdout/stdin close (#1439)
We are seeing some cases for JetBrains team where the Cody agent process
is still running after IntelliJ exits. I'm not sure if the diff in this
PR fixes the problem but I figured it was worth a shot.

I'm open for suggestions to alternative fixes. We have already made
changes on the JetBrains side that fix most of the instances of this
problem. Since those changes, we have only been able to reproduce this
issue with a local sandbox of IntelliJ (not normal installation)

## Test plan

n/a

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->