Fixup: Only apply formatting changes to a range, and clean the undo stack (#1445)
## Description

This PR:
- Only applies the formatting changes to the fixup selection. By calling
the format provider directly and filtering the edit down to the
selection range. VS Code only provides `formatDocument` and
`formatSelection` direct commands, so this is the approach that gives us
the control we need.
- Preserves a single undo/redo action for both the fixup and the
formatting change. A user would have to undo a fixup twice if it also
applied formatting. This should be an implementation detail and not
visible to the user. Now both changes are consolidated into a single
undo/redo action.



https://github.com/sourcegraph/cody/assets/9516420/7f5d6b9f-7c2b-4594-b1ad-25ac2d0b575b



## Test plan

Tested locally, see demo

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->