refactor: update preamble rules and mixin (#1442)
Updating current preamble, rules, and prompt mixin with new findings
from https://github.com/sourcegraph/cody/pull/907.

Issue 1: Cody hallucinates about file path and code

One of the main issues we saw in Chat is Cody hallucinating about files
and code that do not exist in the shared context.

I noticed this is mainly because of how we phrased our context, as Cody
do not understand the relationship between each code snippet we share
with them. For more information, see
https://github.com/sourcegraph/cody/pull/907#discussion_r1320490212

One of the many prompts I've tried to prevent hallucination is to add
`Do not make any assumptions about the code and file names or any
misleading information` to the question, which has been working well for
me. This is also the prompt we added to all our commands. [A customer
has also
confirmed](https://sourcegraph.slack.com/archives/C04MSD3DP5L/p1697747191354829?thread_ts=1697626152.033629&cid=C04MSD3DP5L)
that adding this to their questions stopped Cody from hallucinating and
received better responses from Chat, this is why I propose adding it to
`PromptMixin`, which will be added to the start of every question
(instead of Preabmle, which is added to the start of the session and
will get left out as the conversation gets longer.)

I don't think this will be 100% hallucination-proof, but it works a lot
better than what we currently have.

#### Before

- lib/share/ui/src/components/CodeBlock is not the correct path

![Screenshot 2023-10-18 at 2 45
49 PM](https://github.com/sourcegraph/cody/assets/68532117/9e8b6438-b37c-4745-a5f5-c6abcc2ceec8)


#### After

-  make well-informed answers

![Screenshot 2023-10-18 at 2 47
07 PM](https://github.com/sourcegraph/cody/assets/68532117/87d8b9ce-b8bc-4298-a7be-9a6957a58c36)

Issue 2: The current preambles need to be updated, and not working as
intended most of the times.

Addressed by this PR with the following changes:
- Simplified the preamble actions and rules text to be more concise.
- Removed redundant rules that will always be false, e.g. Cody never has
direct access to your file or repository even when we tell them they do.
- Removed detailed instructions about code formatting and limitations,
leaving just the essential rules
- Shortened the preamble answer to focus on the core assistant persona
and capabilities

Issue 3: Cody will not answer in language that is not the default
language of the editor

CLOSE https://github.com/sourcegraph/cody/discussions/1011 &&
https://github.com/sourcegraph/cody/issues/988

Many users have complained that they cannot get Cody to answer their
questions because Cody refused to answer questions that is not the same
as the default language of the editor.

Addressed by this PR with the following changes: Remove
`languagePromptMixin` from the `activate` event for editors, which
allows Cody to answer in the same language as the question

#### Before

<img width="609" alt="Screenshot 2023-10-19 at 5 08 17 PM"
src="https://github.com/sourcegraph/cody/assets/68532117/6265a46e-d5e8-489b-a92d-824b56a391dc">

#### After

<img width="624" alt="Screenshot 2023-10-19 at 5 07 57 PM"
src="https://github.com/sourcegraph/cody/assets/68532117/b1478a0f-8bfa-468b-abe2-69588f1deef6">

## Test plan

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

1. Ask Cody a question in another language that is not the same as your
editor. Cody should still be able to answer your question in the same
language.
2. Ask Cody a question in editor, and then ask the same question in Web
to compare the quality. Questions where Cody hallucinates on Web should
not happen in the editor when using the updated prompts.

Example question to ask in the cody repo: `Do we format the completion
item before we suggest them?`

<img width="883" alt="image"
src="https://github.com/sourcegraph/cody/assets/68532117/86b1c8c5-4e2c-4674-9ad6-18382dbb2813">