Autocomplete: Fix suggest item race condition (#1477)
This PR fixes a race condition that was observed when picking items from
the suggest widget. The main reason for this was that the
`reuseLastCandidate` logic was not aware of the _actual_ document state
but only comparing the doc context `prefix`. This, in some cases, can
have the suggest widget candidate injected.

Now, for that logic, these two completions are exactly the same:

- A completion request for `console.█` with `log` selected in the
suggest widget
- A completion request for `console.log█`

This made this logic conclude that it is safe to return the same suggest
(`log`) in both cases, causing the repitition.

While working on this, I also noticed another issue with how we
calculate the `current-doc`: Right now, this would only work if the
selection to be replaced is non-emtpy. This seems to always be the case
in production (thankfully) since suggest widget will only trigger once
you insert a `.` and it will always replace this dot (so in the above
case, it will be `console.█` with `.log` selected, with a range that
replaced the `.`). This is not how we wrote our unit tests though and
conceptually it shouldn't require this. Hence this PR also adds support
for empty ranges in the current doc logic together with a bunch of
tests.

This was likely introduced by us starting to propagate the
`completeSuggestWidgetSelection` flag properly (which before was always
false) but even before that, this race condition could sometimes happen.
This implementation should be a bit more robust.

## Test plan

### Before


https://github.com/sourcegraph/cody/assets/458591/df0cf31e-113a-41b2-a58d-9da9709b29a5

### After


https://github.com/sourcegraph/cody/assets/458591/9071bb54-1df9-4cca-9cbd-b8e307768fa4





<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

---------

Co-authored-by: Beatrix <68532117+abeatrix@users.noreply.github.com>