Fix version numbering script (#51)
Previously, the patch number was computed based on the release timestamp
of the last Sourcegraph release. This was problematic because it meant
we reset the patch number for every new patch release of Sourcegraph.
This PR fixes the problem by making the patch number be based on the
release timestamp of the latest MAJOR.MINOR.0 release.


## Test plan
```
# This PR
❯ ./scripts/next-release.sh
5.2.3528
# On main
❯ ./scripts/next-release.sh
5.2.841
```

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles

Why does it matter? 

These test plans are there to demonstrate that are following industry
standards which are important or critical for our customers.
They might be read by customers or an auditor. There are meant be simple
and easy to read. Simply explain what you did to ensure
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component: 
- "Covered by existing tests" or "CI" for the shortest possible plan if
there is zero ambiguity
  - "Added new tests" 
- "Manually tested" (if non trivial, share some output, logs, or
screenshot)
- Updating docs: 
  - "previewed locally" 
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if
incorrect
  - "CI" 
  - "locally tested" 
-->