Fix Apple M1 issue for CI releases (#55)
The CI release workflow was missing a dependency on QEMU and binfmt to
correctly sign the agent-macos-arm64 binary.

Massive kudos to @burmudar for the help with this!
https://github.com/sourcegraph/cody/pull/1474

## Test plan

Merge this PR, cut a new nightly release and manually verify that it
works on Apple M1.
<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles

Why does it matter? 

These test plans are there to demonstrate that are following industry
standards which are important or critical for our customers.
They might be read by customers or an auditor. There are meant be simple
and easy to read. Simply explain what you did to ensure
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component: 
- "Covered by existing tests" or "CI" for the shortest possible plan if
there is zero ambiguity
  - "Added new tests" 
- "Manually tested" (if non trivial, share some output, logs, or
screenshot)
- Updating docs: 
  - "previewed locally" 
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if
incorrect
  - "CI" 
  - "locally tested" 
-->