Fix task-controller e2e test flakes on Windows
On Windows without this change, I always see this failure in task-controller e2e tests:

```
    Error: expect(received).toEqual(expected) // deep equality

    - Expected  - 1
    + Received  + 0

      Array [
    -   "cody.auth/failed",
        "cody.auth/connected",
        "cody.command.edit/executed",
        "cody.recipe.fixup/executed",
        "cody.fixup.apply/succeeded",
      ]
```

The `"cody.auth/failed"` event is never logged. I'm not too familiar with this, but I think this is because the mock server is not up and running at the time the extension activates and first tries to sync authentication.

This change starts the server up slightly earlier and appears to avoid the flake. Reducing the timeout on line 69 also resolves it for me locally, but that feels like it might re-introduce other issues it was designed to fix.