Autocomplete: do adjust a completion range if it does not match the current line suffix (#1507)
- Fixes the issue where completions were needlessly hidden:
`myFunction(█)` with the insert text being `"one", false` because the
current line suffix `)` was not present in the generated completion.
However, we still adjusted the completion insertion range to include it.
- Fixes the issue where all the generated completions were hidden, even
if only one needed to be ignored. Now, we filter out invisible
completions individually and keep visible completions if we have any.