Edit: Fix recursive diff expansion (#1621)
## Description

We use a temp document to provide a "before" version of the edited fixup
file. There is an edge case where, if we are adding new code (thus
deleting from the temp document), we can reuse the already diffed temp
document. This leads to the diff expanding every time it is triggered,
as we delete more code from the selection.

We now ensure each diff has its own unique ID, so it always uses a fresh
"temp" document.

### Before


https://github.com/sourcegraph/cody/assets/9516420/c584f8a5-e21c-468f-bb82-c9eece29cc38


### After


https://github.com/sourcegraph/cody/assets/9516420/20eb4fd5-b318-47d2-b66c-c9a3b9d076c1


## Test plan

Tested locally:
- Diff for edits that expand ranges
- Diff for an edit where the doc has been changed elsewhere after it was
applied

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->