Autocomplete: Add feature flag for Sourcegraph specific Fireworks setup (#1670)
Now that I have added support for
`fireworks/accounts/sourcegraph/models/starcoder-16b` and
`fireworks/accounts/sourcegraph/models/starcoder-7b` to the Cody
Gateway, we can start using it on the dotcom clients.

To roll this out, I’m adding a new flag
`cody-autocomplete-default-starcoder-hybrid-sourcegraph` that is only
used when `cody-autocomplete-default-starcoder-hybrid` is also true.
This way we can set
`cody-autocomplete-default-starcoder-hybrid-sourcegraph` to a 50:50
split and it will only affect the test group that is already in the
StarCoder branch.

I expect this to be very short lived but it would be great if we can
measure the affect of that migration.

I plan to roll this out as a patch release so we can turn this on in the
next days.

<img width="996" alt="Screenshot 2023-11-07 at 16 12 05"
src="https://github.com/sourcegraph/cody/assets/458591/2ebd0511-bc18-4d7a-bdf5-0ab285772a02">

<img width="1162" alt="Screenshot 2023-11-07 at 16 13 00"
src="https://github.com/sourcegraph/cody/assets/458591/7177c14d-f456-4c46-b2f6-198a0af99b39">

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->