BFG: correctly load git repositories (#1728)
Previously, we tried to infer the git repository based on open
documents. This logic was brittle causing BFG to not pick up
repositories when it should. This PR replaces this logic with direct
calls to the git extension, which gives us access to this information
directly.


## Test plan

See updated `BfgContextFetcher.test.ts`. I also manually tested that BFG
is loaded when running the new autocomplete evaluation.

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->