Autocomplete: Implement top-k retrieval mixing and make section history a retriever (#1705)
Closes #1641
Cloes #1642

Some more retrieval improvements!

- Section history now implements the retrieval strategy and is now
longer called transitively by the lsp light retriever. This will allow
us to mix it with other results soon.
- The context mixer now does top-k mixing (not sure if this is the right
terminology?): If more than one retrieval strategy is used, it will pick
the top most results from every strategy and then follow on with the
second rank, etc. For now this is only used in the `lsp-light` strategy
that isn't used but the idea is that it will contain the highest ranked
lsp light result followed by the highest ranked section history and then
the second highest ranked lsp light etc.
- This is only temporary and we'll add more advanced strategies shortly.
Again this only affects `lsp-light` strategy which is not rolled out
right now.
- Overhauls the `contextSummary` logging to include a ton more
information like:
  - Execution time of each retrieval strategy
- Number of retrieved items and number of items included in the
suggested context set
- A bitmap that can be used to know which types of snippets are
included, more on this below.
- Total number of characters suggested for the request context (this may
NOT be the exact prompt length but it can be used as a proxy for it)

## Position bitmap

Okay maybe this is over engineering but I do think we need to understand
how the top snippets are composed when we start to experiment with
snippet mixing. To do that, we can use bitmaps and setting the bit
according to the position in the result set for every type to express
the order.

Here's an example of retrieval stats for an `lsp-light` request that
contains a total of 5 items.

```json
{
  "lsp-light":{
    "suggestedItems":4,
    "positionBitmap":29,
    "retrievedItems":4,
    "duration":123
  },
  "section-history":{
    "suggestedItems":1,
    "positionBitmap":2,
    "retrievedItems":1,
    "duration":123
  }
}
```
 
In binary, the `positionBitmap` for each strategy look like this:

- lsp-light: `11101`
- section-history: `00010`

Since the result is read from the lowest bit to the highest, we can see
that the first item (right most bit) is from the lsp light set, the
second entry from the section history and the remaining three from the
lsp light set again.

## Test plan

### Jaccard similliarty still works

<img width="1468" alt="Screenshot 2023-11-09 at 17 08 40"
src="https://github.com/sourcegraph/cody/assets/458591/76d30fd3-cd7f-47c6-8a1f-54e4fdf41652">

### lsp light works

<img width="2053" alt="Screenshot 2023-11-09 at 18 12 56"
src="https://github.com/sourcegraph/cody/assets/458591/26782928-a93f-42b4-8e0a-36320fa55ed3">


<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->