Autocomplete: Implement a variant of RRF for retrieval mixing (#1710)
Closes #1643

Building on top of #1705, this PR adds a varient of reciprocal rank
fusion when fusing multiple retrieval sets.

Our big problem though is that we do not have a stable concept of a
document between retrieval strategies. The results are arbitrary ranges
within files and when combining two retrieval strategies with each
other, the ranges can be different (e.g. semantic for BFG and fixed
windows for jaccard similiarty).

To overcome this limitation, we define the whole file as a document and
rank the files based on RRF. This creates two limitations:

1. Since a retrieval strategy can return multiple ranges of the same
file, we need to ensure we only pick the top rank for a document inside
the retrieval set.
2. When fusing the results, we only know the order of documents that we
want to include, not the order of ranges inside this document. For now,
we use top-k sorting for that (so we include the top most entries for
each retrieval strategy until we exhaust our character budget).
   
This means that we start by exhausting _all_ snippets for the top ranked
file and before we move on to snippets from the second file. We'll have
to see how this behaves in practice and if we need to update our
definition of a document again.

This PR also refactors the context strategy into a factory class which
we can easily mock in tests.

## Test plan

Added a test case.

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->