fix: handle @ files with start line 0 (#1734)
The file link markdown generation was previously erroring when the start
line was 0. This fixes it by checking for a valid end line instead of
start line.

Also simplifies the regex used to replace the file name with a markdown
link.

## Test plan

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

Added new test for edge case error when generating markdown links for
selections starting at line 0.