Agent: support multiple workspaces with `evaluate-autocomplete`
Previously, `evaluate-autocomplete` required all the arguments to be
passed from the command-line and it was only able to evaluate a single
workspace at a time (with no parallelization!). Now, you can evaluate
multiple workspaces in parallel based on JSON configuration inside a
file. This makes it easy to run evaluations on a cron job to track
progress.