BFG: rename `bfg` binary name into `cody-engine` (#1736)
The name of the binary appears in activity monitors and we want to make
it easy for users to trace it back to Cody. We still keep the "bfg"
codename in the implementation (variable names, JSON-RPC protocol, etc.)
  because it requires non-trivial engineering effort to change the name
  everywhere and we don't think this is worth the effort at this early
  stage of the project.

I kept the `graphContext: 'bfg'` setting name because that is
specifically about code graph. I'm open to changing that as well to
`cody-engine` but I don't think it's necessary.

## Test plan

Ran extension manually and confirmed that the output channel doesn't
print BFG-related messages. Also confirmed that the binary name is now
named "cody-agent-*"
<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->