Add config error handler (#1739)
This PR closes #1667. 
I have implemented a function that checks for the allowed enum values
for the Cody config. If the config value is not from the allowed list of
enum values, then it shows an error during startup.


## Test plan

Open the `settings.json` and change
`cody.autocomplete.advanced.provider` to `gpt-4`. After this, reload the
window a VSCode error message will be displayed during the startup.

---------

Co-authored-by: Philipp Spiess <hello@philippspiess.com>