Autocomplete: split doc context calc logic (#1784)
- This PR splits the `docContext` calculation logic into multiple
smaller functions to clarify the dependencies between its fields. It's
required for #1619 because I plan to update the existing doc context by
adding the first line of completions to the prefix and re-calculate all
the other doc context fields afterward. To avoid creating a fake
document with the updated text, we now have a separate
`getDerivedDocContext` function, which accepts a pre-calculated prefix
and suffix, and returns the document context.
- It also removes a couple of `document` dependent fields that are easy
to calculate on the spot from the `docContext`.
- No functional changes.
- Planned follow-up for a separate PR (to complete the removal of the
`syntactic-triggers` feature flag). #1783