evaluate-autocomplete: fix git-log strategy and add more CSV columns (#1859)
Previously, the `git-log` strategy did not run correctly due to a
regression in the last change. This PR fixes that regression and adds
two new columns to the generated CSV file:

- `EVENT`: for detailed information about each autocomplete response
like latency. This is the JSON object we include in telemetry events.
- `REVISION`: to be able to know at which git commit the request was
sent. This is mostly helpful to know if two requests for the same
file/line location happened on the same file content.


## Test plan

Manually tested.
<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->