Agent: add testing mode
Previously, we didn't run agent tests in CI because they relied on a
live Sourcegraph server running in the background to respond to HTTP
requests. This PR adds support for a new testing mode that records HTTP
requests into a directory and replays from that directory. This new
testing mode makes it possible to enable the agent tests in CI because
they run entirely from the HTTP recording avoiding the need for a
sourcegraph access token or a live sourcegraph server. The HTTP recorder
has been customized to automatically redact the Sourcegraph access token
so that we don't commit the token into the git repository.