Simplify chat prompt construction (#1717)
This is a work-in-progress branch that aims to rebuild the prompt
construction process for chat from scratch.

Why the need for a rewrite? The existing prompt construction process
goes through a bunch of different components that were built for a
previous version of chat that was more complex. The same code was also
used as the context + prompt construction for recipes and commands.
There seem also to be old assumptions baked into the code, such as only
one chat at a time, which seems to be addressed by mutating shared
state. I encountered this complexity when trying to add symf as a new
context provider.

Rethinking from scratch, chat should be very simple. Each chat panel is
logically independent. For each, we need to maintain:
* The actual transcript of human messages and assistant responses
* A list of context items (currently doc and source code file snippets)
* A prompt constructor that takes in the transcript and the context
items and returns the request that is actually sent to the LLM
* The toggles in the chat panel that affect the backend behavior (e.g.,
which LLM is selected)

This branch show contains a proof of concept. It is runnable and the new
behavior can be toggled on with the
`cody.experimental.simpleChatContext` setting:


https://github.com/sourcegraph/cody/assets/1646931/cb2cda9e-68ef-4c91-affa-e1195ac62429

## Code overview

* [The code that swaps the old `ChatPanelProvider` controller class
instance for the new `SimpleChatPanelProvider`
instance](https://github.com/sourcegraph/cody/pull/1717/files#diff-dca82f1d9c5f3cef4ad40e19254d228e349b29db3ef93a862283b06032ccbd39R122-R149).
The `useSimpleChatPanelProvider` field toggles whether the new class is
used.
*
[`SimpleChatPanelProvider`](https://github.com/sourcegraph/cody/pull/1717/files#diff-373d17df4ed754b63519de6eb54fd039776998abccf5fcaac21a47e04a53349eR1)
is the new controller class. Its implementation is incomplete, but it
provides enough functionality to do the basic call-and-response with
context fetching
* The context fetching is done in the
[`onHumanMessageSubmitted`](https://github.com/sourcegraph/cody/pull/1717/files#diff-373d17df4ed754b63519de6eb54fd039776998abccf5fcaac21a47e04a53349eR225)
method. Right now, it just fetches 2 doc and 2 source snippets from the
embeddings backend. This should be updated to mimic what's currently
done (i.e., preserve the old behavior).
* The class also needs to be filled in to support the other behavior in
the new chat panel.
*
[`SimpleChatModel`](https://github.com/sourcegraph/cody/pull/1717/files#diff-a1cc9f50c75dcc85fc841ac2c97f15e4b7abeb27de1f909cd9b59d46ca70afa1R1)
is the model class for the chat. Currently, it just keeps track of
messages and context items.

## Test plan

Set `"cody.experimental.simpleChatContext": true` and run through the
following:

- [ ] Test basic chat creation (create a new panel and ask a trivial
question)
- [ ] Ask a question about the codebase and verify relevant context is
fetched and used
- [ ] Test multiple follow-up questions
- [ ] Test asking questions until context window is exhausted. (You can
manually change the max context window in the code to avoid the need to
have to generate lots of text.)
- [ ] Test @-mentioning specific files and ensure the context of these
files is used
- [ ] Test edit
- [ ] Test LLM selector - select GPT-4 and ask Cody to provide its name.
- [ ] Test message restore and save
- [ ] Create a message with the old chat panel, save it, and then open
it with the new chat panel
  - [ ] Save a message with the new chat panel. Restore it
- [ ] Test reset chat button and start a new chat from the empty state
- [ ] Check that context files are linked to selected ranges in files
- [ ] Test insert and copy code buttons
- [ ] Test "enhanced context" checkbox - uncheck and verify enhanced
context is *not* fetched

### Not yet implemented

The following features are not yet implemented and will be implemented
in a follow-up PR (which will also remove the
`cody.experimental.simpleChatContext` flag.
- [ ] Support for `explain` command
- [ ] Support for `smell` command
- [ ] Support for custom commands
- [ ] Include "You are Cody" text in the preamble
- [ ] Input history (up and down arrow keys should scroll through
previous user inputs in the chat input box)

### Bug list (to fix before merge)
- [ ] Clicking on context file link only opens file, but not selected
range
- [ ] @-mentioned files show up as absolute URIs and aren't opened when
clicked